requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-filter==23.3
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
